fix(runtime,hir): Function parity v2 round 9-11 — proxy ToString/ToPrimitive + js_is_symbol floor, apply(this,arguments) raw argArray + arguments-object unpack, arg-expression const-eval (p+","+p), fn-decl .constructor kinds, PutValue caller poison, class-ref stringification
